Darkly theme modified to mimic "Amethyst" from VS
A few theme changes were made to the main Darkly UI in order to try and mimic the "Amethyst" theme pack from Visual Studio. These changes were added to the main CSS assets file in order to avoid breaking the index page...for now 🗡️ 

Any modifications here are simple and not meant to drastically alter the state of the site. More to come!